test-requirements.txt
odoo-addon-edi_oca @ git+https://github.com/thienvh332/edi-framework.git@refs/pull/6/head#subdirectory=edi_oca
